Reorganize Channel components for clarity
