+++ nextjs/components/Pages/Channel/Content/ChannelGrid/index.module.scss
