+++ nextjs/components/Pages/Channel/Content/ChannelGrid/index.tsx
@@ -1,5 +1,5 @@
-import CustomLink from '../../Link/CustomLink';
-import { SerializedThread } from '../../../serializers/thread';
+import CustomLink from '../../../../Link/CustomLink';
+import { SerializedThread } from '../../../../../serializers/thread';
 import ChannelRow from '../ChannelRow';
 import type { Settings } from 'serializers/account/settings';
 import { SerializedUser } from 'serializers/user';
