+++ nextjs/components/Pages/Channel/Content/ChannelRow/index.module.scss
