+++ nextjs/components/Pages/Channel/Content/PinnedThread/index.module.scss
