+++ nextjs/components/Pages/Channel/Content/PinnedThread/index.test.tsx
