+++ nextjs/components/Pages/Channel/Content/PinnedThread/index.tsx
