+++ nextjs/pages/d/[communityName]/c/[channelName]/index.tsx
@@ -1,5 +1,5 @@
 import { channelGetServerSideProps } from 'services/channel';
-import Channel from 'components/Pages/ChannelsPage';
+import Channel from 'components/Pages/Channel';
 import { GetServerSidePropsContext } from 'next/types';
 
 export default Channel;
